moved time visualization code
