+++ src/js/components/account/visualizations/time/TimeVisualization.jsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import ChartMessage from 'components/search/visualizations/time/TimeVisualizationChartMessage';
+import ChartMessage from 'components/sharedComponents/timeChart/TimeVisualizationChartMessage';
 import BarChartStacked from './chart/BarChartStacked';
 import TimeTooltip from './TimeTooltip';
