+++ src/js/components/sharedComponents/timeChart/chart/BarChart.jsx
