+++ src/js/components/sharedComponents/timeChart/chart/BarChartLegend.jsx
