+++ src/js/components/sharedComponents/timeChart/chart/BarChartLegendItem.jsx
