+++ src/js/components/sharedComponents/timeChart/chart/BarItem.jsx
