+++ src/js/components/sharedComponents/timeChart/chart/BarXAxis.jsx
