+++ src/js/components/sharedComponents/timeChart/chart/BarXAxisItem.jsx
